requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.2.17
django-allauth==0.50.0
django-filter==24.3
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
oauthlib==3.2.2
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
django-cors-headers
channels